Fixed GetAxis Error preventing Menu scene from Running
Applied changes to Input Manger and player movement scripts so that the original "Horizontal" axis may be recognized properly in other scenes.
Error was happening because "Horizontal" had been renamed to "Horizontal1" and so it wasn't being recognized else place (i.e Main Menu scene) where the original "Horizontal" Axis was attempted to to be accessed by Unity native scripts.